// src/app/model/plantillas.ts
export interface Plantillas {
  id: number;
  codigo: string;
  nombre: string;
  destinatario: string;
  copia: string;
  contenido: string;
  estado: string;
  asunto: string;
}

export interface PlantillaReybanpac {
  id: number;
  titulo: string;
  contenido_video: string;
  contenido: string;
  estado: string;
}